pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sec-mcp"
version = "0.1.0"
description = "MCP server for SEC filing analysis with BERT-based NLP"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "fastmcp>=2.0",
    "edgartools>=3.0",
    "transformers>=4.40",
    "torch>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sentencepiece",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: requires network access to SEC EDGAR",
    "slow: involves model loading",
]

[tool.ruff]
target-version = "py310"
line-length = 100
